requirements.txt
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-filter==25.2
django-cors-headers==4.9.0
python-decouple==3.8
Pillow==12.1.1
psycopg2-binary==2.9.11
celery==5.6.3
redis==7.4.0
django-celery-beat==2.9.0
django-celery-results==2.6.0
elastic-apm==6.23.0
psutil==7.0.0
python-dateutil==2.9.0.post0
